requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
WTForms==3.0.1
psycopg2-binary==2.9.6
pytest==7.4.0
coverage==7.3.1
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
python-multipart==0.0.6